practice/shred_recovery/requirements.txt
numpy==1.14.5
Pillow==5.2.0
